Use "set" instead of "splice" for child changed
When using firebase-collection in a dom-repeat, the whole template has to refresh every time a child changes instead of only the modified child. That's because it is updating the array with a splice.

If we use a set instead, only the actual child is refreshed. The main problem with the "splice" was that I was losing focus on an input when the refresh was happening.